rename packages and namespaces remove boilerplate codes to create out file
