+++ .travis.yml
@@ -6,5 +6,5 @@ cache:
     - $HOME/.sbt/root
 before_script:
   - sbt jmh:compile
-  - sbt avroBench/run circeBench/run csvBench/run jsoniterScalaBench/run protoBench/run thriftBench/run
+  - sbt avroBench/run circeBench/run csvBench/run jsoniterScalaBench/run msgpackBench/run protoBench/run thriftBench/run
   - sbt avroBench/jmh:run circeBench/jmh:run csvBench/jmh:run jsoniterScalaBench/jmh:run protoBench/jmh:run thriftBench/jmh:run
