+++ avro-bench/src/main/scala/com/github/saint1991/serialization/benchmark/avro/FileGen.scala
@@ -1,4 +1,4 @@
-package com.github.saint1991.serialization.benchmark
+package com.github.saint1991.serialization.benchmark.avro
 
 import java.io.File
 
@@ -8,23 +8,21 @@ import org.apache.avro.Schema
 import org.apache.avro.file.DataFileWriter
 import org.apache.avro.specific.SpecificDatumWriter
 
+import com.github.saint1991.serialization.benchmark.FileUtil
+
 object FileGen extends App {
 
   final val Schema = Nobid.SCHEMA$
   val writer = new SpecificDatumWriter[Nobid](Schema)
 
-  final val N = 10
+  final val N = 10000
   val dataset = DataSet.createDataset(N)
 
-  final val out = new File("out")
-  if (!out.exists()) out.mkdir()
-  final val outFile = new File("out/nobids.avro")
-  outFile.createNewFile()
-
+  val outFile = FileUtil.mkOutFile("nobid.avro")
   val outFileWriter = new DataFileWriter[Nobid](writer)
 
   // write to file
-  writeToFile(dataset, Schema, outFile, outFileWriter)
+  writeToFile(dataset, Schema, outFile.toJava, outFileWriter)
 
   private def writeToFile(dataset: Seq[Nobid], writerSchema: Schema, file: File, writer: DataFileWriter[Nobid]): Unit = allCatch andFinally {
     writer.flush()
