+++ csv-bench/src/test/scala/com/github/saint1991/serialization/benchmark/csv/CsvBench.scala
@@ -1,12 +1,7 @@
-package com.github.saint1991.serialization.benchmark
+package com.github.saint1991.serialization.benchmark.csv
 
-import java.nio.charset.StandardCharsets
 import java.util.concurrent.TimeUnit
 
-import io.circe._
-import io.circe.syntax._
-import io.circe.generic.auto._
-import io.circe.parser._
 import org.openjdk.jmh.annotations.{BenchmarkMode, Fork, Measurement, Mode, OutputTimeUnit, Scope, State, Warmup, Benchmark => JmhBenchmark}
 
 import com.github.saint1991.serialization.benchmark.dataset._
@@ -27,24 +22,22 @@ import com.github.saint1991.serialization.benchmark.dataset._
   "-XX:+AlwaysPreTouch"
 ))
 @OutputTimeUnit(TimeUnit.SECONDS)
-class CirceBench {
+class CsvBench {
+  import com.github.saint1991.serialization.benchmark.csv.Csv._
+
   final val N = 100000
   val dataset: Seq[Nobid] = DataSet.createDataset(N)
 
-  implicit val encoder: Encoder[SpotType.Value] = Encoder.enumEncoder(SpotType)
-  implicit val decoder: Decoder[SpotType.Value] = Decoder.enumDecoder(SpotType)
-
-  val encodedDataset: Seq[Array[Byte]] = encode()
+  val encodedDataset: Seq[String] = encode()
   decode()
 
   @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
-  def encode(): Seq[Array[Byte]] = {
-    dataset.map(_.asJson.noSpaces.getBytes(StandardCharsets.UTF_8))
+  def encode(): Seq[String] = {
+    dataset.map(toCsv)
   }
 
   @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
   def decode(): Seq[Nobid] = {
-    encodedDataset.map(str => parse(new String(str, StandardCharsets.UTF_8)).right.get.as[Nobid].right.get)
+    encodedDataset.map(fromCsv)
   }
 }
-
