+++ jsoniter-scala-bench/src/test/scala/com/github/saint1991/serialization/benchmark/jsoniter/JsonIterScalaBench.scala
@@ -1,10 +1,12 @@
-package com.github.saint1991.serialization.benchmark
+package com.github.saint1991.serialization.benchmark.jsoniter
 
 import java.util.concurrent.TimeUnit
 
+import com.github.plokhotnyuk.jsoniter_scala.core._
+import com.github.plokhotnyuk.jsoniter_scala.macros._
 import org.openjdk.jmh.annotations.{BenchmarkMode, Fork, Measurement, Mode, OutputTimeUnit, Scope, State, Warmup, Benchmark => JmhBenchmark}
 
-import com.github.saint1991.serialization.benchmark.nobid.Nobid
+import com.github.saint1991.serialization.benchmark.dataset._
 
 @State(Scope.Thread)
 @Warmup(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
@@ -22,17 +24,19 @@ import com.github.saint1991.serialization.benchmark.nobid.Nobid
   "-XX:+AlwaysPreTouch"
 ))
 @OutputTimeUnit(TimeUnit.SECONDS)
-class ProtoBench {
-
+class JsonIterScalaBench {
   final val N = 100000
   val dataset: Seq[Nobid] = DataSet.createDataset(N)
+
+  implicit val codec: JsonValueCodec[Nobid] = JsonCodecMaker.make[Nobid](CodecMakerConfig())
 
   val encodedDataset: Seq[Array[Byte]] = encode()
   decode()
 
   @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
-  def encode(): Seq[Array[Byte]] = dataset.map(_.toByteArray)
+  def encode(): Seq[Array[Byte]] = dataset.map(x => writeToArray(x))
 
   @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
-  def decode(): Seq[Nobid] = encodedDataset.map(Nobid.parseFrom)
+  def decode(): Seq[Nobid] = encodedDataset.map(str => readFromArray(str))
 }
+
