+++ msgpack-bench/src/main/scala/serialization/benchmark/msgpack/FileGen.scala
@@ -1,0 +1,5 @@
+package serialization.benchmark.msgpack
+
+object FileGen extends App {
+  // TODO
+}
