+++ proto-bench/src/main/scala/com/github/saint1991/serialization/benchmark/protobuf/FileGen.scala
@@ -1,22 +1,20 @@
-package com.github.saint1991.serialization.benchmark
+package com.github.saint1991.serialization.benchmark.protobuf
 
-import java.io.{File, FileOutputStream, OutputStream}
+import java.io.{FileOutputStream, OutputStream}
 
 import scala.util.control.Exception._
 
-import com.github.saint1991.serialization.benchmark.nobid.Nobid
+import com.github.saint1991.serialization.benchmark.FileUtil
+import com.github.saint1991.serialization.benchmark.protobuf.nobid.Nobid
 
 object FileGen extends App {
 
   final val N = 10000
   val dataset = DataSet.createDataset(N)
 
-  final val outDir = new File("out")
-  if (!outDir.exists()) outDir.mkdir()
-  final val outFile = new File("out/nobids.protobuf")
-  outFile.createNewFile()
+  final val outFile = FileUtil.mkOutFile("nobid.protobuf")
 
-  val out = new FileOutputStream(outFile)
+  val out = new FileOutputStream(outFile.toJava)
 
   // write to file
   allCatch andFinally {
